fix proj

+++ src/common/libs/include/common/calibration.hpp
@@ -32,9 +32,8 @@ static Eigen::MatrixXd proj(const Eigen::Matrix3d& K, const Eigen::Matrix3d& R, 
     std::vector<cv::Point3f> points_lidar(x.cols());
     for (int i = 0; i < x.cols(); i++)
         points_lidar[i] = cv::Point3f(x(0,i),x(1,i),x(2,i));
-
     // Project the points
-    std::vector<cv::Point2d> points_distorted_cv;
+    std::vector<cv::Point2f> points_distorted_cv;
 
     cv::Mat R_mat(3,3,CV_64F);
     cv::eigen2cv(R,R_mat);
@@ -49,6 +48,7 @@ static Eigen::MatrixXd proj(const Eigen::Matrix3d& K, const Eigen::Matrix3d& R, 
 
 
     cv::projectPoints(points_lidar, R_mat, t_mat, K_mat, D_C_cv, points_distorted_cv);
+    
 
     // std::cout << "Distort complete" << std::endl;
 
@@ -59,6 +59,8 @@ static Eigen::MatrixXd proj(const Eigen::Matrix3d& K, const Eigen::Matrix3d& R, 
     {
         points_distorted(0,i) = points_distorted_cv[i].x;
         points_distorted(1,i) = points_distorted_cv[i].y;
+        // if (i%20 == 0)
+        //     std::cout << "Point: " <<  points_distorted(0,i) << ", " <<  points_distorted(0,i) << std::endl;
     }
